Immutable List and Set
This patch implements immutable versions of List and Set. fastutil
already had helper methods like "IntLists.unmodifiable(list1)" but
the resturned list in this case was just an unmodifiable view of
list1 and not truely immutable. Updating list1 would still result
in changes in the unmodifiable view. Also the unmodifiable wrapper
meant that every access to the underlying list had to go through
two pointers instead of one. It also had the additional memory
overhead of the wrapped object.

The ImmutableList and ImmutableSet implementations here use Builder
pattern or static methods to collect the data in the collections and
then build the immutable collection from that. Hence in addition to
immutability, they also use the minimum memory required for the
provided size of data.