Don't rewrap already correct type Consumers
calling forEachRemaining(action) where action's compile time type was
Consumer<X> but it's runtime type was XConsumer (the primitive consumer
type) was causing a rewrap in a new lambda into being a primitive
consumer when it already was an primitive consumer.

This has a _substantial_ speedup when working with the primitive stream,
by a factor of 10 in some cases.