Make every unit test and main test use unique filename
Previously, every test tried to serialize to the same file,
"it.unimi.dsi.fastutil.test". This mean if the tests were run in
parallel, they would "step on each others toes" and overwrite each
others' serialized files, causing strange exceptions in the
deserialization tests when run in parallel.

Now each test has a unique file it writes to by keying off the name of
the class of the thing being serialized.