Add a type-specific Spliterator that wraps iterators
This allows up to have a default implementation of spliterator that does
not rely on widened iterators. This means that the horrible hack of the
ImplementsSpliterator interface is no longer needed, and has been
excised.

The cost for this is yet more classes in Spliterators and no longer
getting potential future performance benefits from the JDK's version of
this method (though that is rather unlikely; there is only so much you
can do wrapping an Iterator)