Fix various small issues with Spliterators.
Fix concating Spliterator not returning Long.MAX_VALUE if size estimate
overflowed.
Fix comment suggesting that trySplit() is always linear time, when it
was intended to mean that it will be for only the returned instance.
Add comment to another place where warning will be raised but why we are
not supressing it.