Validate that duplicate entries to the "of" method of set throws.
This brings it into line with what the platform does for these
of methods of Sets.

Because of this validation, it is now worth it to have overloads
of 0 and 1 element, which skip the temporary array and validation
(0 and 1 element sets can never have duplication).

Also in the spirit of ArraySet's "just trust me" API of working
with raw arrays unvalidated, an it has an offshoot "ofUnchecked"
that takes the elements as is and does no validation, like the
array accepting constructors do.