Replace KEY_CLASS_Reference with KEYS_USE_REFERENCE_EQUALITY
This is much clearer that we mean the types that use `==` for equality
testing instead of `.equals`, and not confuse it with `KEYS_REFERENCE`
macro which means any types using Object backed members.

Also documents the gotcha of trying to use a Reference based and a
non-Reference based type together (e.g. you may get non symmetric `equals`)

Also does the same for VALUE.