Partial revert of commit 7615e085f5478f79e8a3edd6dcb34088326a9101
Turns out, in the Java API and fastutil, there are no Itrator backed Spliterators
that actually gives the iterator to multiple splits, and thus the iterator will
never be used by multiple threads. This makes the iterator's thread safety, or
lack thereof, is a non-issue.

The overrides of spliterator where useful though, so those have been kept, just
now using fastiterator if possible.