Fix AbstractMap containing a stack overflow containsValue
If both the default `containsValue` and the default `values` methods are
inherited by an implementating class, they will get a `StackOverflowError`
on trying to test for a value being present.

This is because the default `values().contains` calls `containsValue`, but
the default `containsValue` calls `values().contains`.

Now make the default `containsValues` loop over the entry set.

Add a comment that you should probably override this is your `values()`
has a better implementation. Same also retroactively applied to `containsKey`
and `keySet()`.

Finally, expands `AbstractInt2IntMapTest` to include the generic suite of tests
from `Int2IntMapGenericTest`, with a dummy testonly implementation that implements
the bare minimum for a mutable Int2Int map, to test as many of the `AbstractInt2IntMap`'s
default methods as possible.