Fix FastBufferedOutputStream.close() exception
If the output is closed before FastBufferedOutputStream is closed, then
close() would throw because flush() would try to write the buffer,
even if there's nothing to write.

This can happen if the output is a socket, for example, that is managed
separately from the output stream.

Fix by not calling write() if there is nothing to write.

If the buffer is not empty, then it would still try to write it, and
would throw, which is probably correct.