Introduced `MovableBidirectionalIterator`
This is a memory optimization for the case when a user required to
modify iterator by moving it to a desire position. This works the same
way as `iteraotr(fromElement)` but doesn't create a new iterator that
decreases memory footprint at an algorithms which makes a lot of jumps.

So, user simple calls `move(fromElement)`.

It also introduces `begin()` which moves iterator to the first position,
and `end()` which moves iterator to the last element.